[feat] #15 package 정리
