+++ app/src/main/java/com/aurora/carevision/feature/admin/home/AdminListScreen.kt
@@ -1,5 +1,5 @@
 // AdminListScreen.kt
-package com.aurora.carevision.feature.admin.auth.list
+package com.aurora.carevision.feature.admin.home
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
