+++ app/src/main/java/com/aurora/carevision/feature/nurse/auth/signup/NurseSignUpInfoScreen.kt
@@ -1,4 +1,4 @@
-package com.aurora.carevision.feature.nurse.auth.login
+package com.aurora.carevision.feature.nurse.auth.signup
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -6,6 +6,8 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -23,7 +25,7 @@ import com.aurora.carevision.core.component.TopAppBarLeft
 @Composable
 fun NurseSignUpInfoScreen() {
     Column(
-        modifier = Modifier.fillMaxSize().background(White)
+        modifier = Modifier.fillMaxSize().background(White).systemBarsPadding().statusBarsPadding()
     ) {
         TopAppBarLeft()
         Text(
